Added dynamic clock divider and test
Allows to dynamically adjust the clock
division based on an input variable.

The input clock can be directly fed
through (x=0), reduced to half (x=1),
a thrid (x=2) etc.

The max. clock division (i.e. the number
of bits for the variable x) can be set as
a module parameter.