Added X-server and PyQt support for simulation app
In order to run simulation with PyQt from within
the docker container (on a windows host machine)
adjustment in the Dockerfile, the devcontainer.json
are needed and several python packages have to
be installed.

Added a test_app.py just to quickly check that
PyQt GUI applications can be displayed.

Tested with Windows11, VcXsrv 64.1.17.2.0 and
Docker Desktop 4.17.1.