Added simulation PyQt application
Application cyclically calls the tick
function of the mimosa model. It
reads the input of the user (too
hot, too loud, tickle, etc.), updates
the model with this information
and displays the output state
(emotion etc.) and graphs of
the energy, stress and pleasure
resources

Migrated to PyQt6

Developed UI with the PyQt6
designer.

Upgraded test_app.py from PyQt5
to PyQt6