Added skeleton for neurotransmitter modules
Each neurotransmitter is handled separately
and consists of a system which itself contains
a very specific regulator regulating the neuro-
transmitter level and its temporal adjustment.

The neurotransmitter_system is a collection of
all the neurotransmitters in the system (facade).

With systemverilog it could be done much
cleaner. However, the toolchain for the
hardening appears not to support sv.