Added an illness system
In order to make it more interesting, the
mimosa can get ill now. If it is ill, neuro-
transmitter levels and hence emotions
and actions can be messed up quite a
bit.

The mimosa gets ill when the illness
counter reaches a certain value and
regenerates when the illness counter
drops below a threshold again. While
ill, the counter counts down, while it
counts up when healthy. The counting
speed depends on the physical state.
If stressed, it gets ill more quickly and
recovers not as fast.