Added logger application for stm32 mcu
A stm32g474 is used as a co-processor. It is
responsible for the clock signal and for
logging all in- and outputs and sending
it via uart to the host computer.

Main idea: It is a deterministic system and
the input is sufficient to simulate the
behaviour with the PyVerilator simulation.
Therefore, the entire simulation can be
run in parallel and we get insight into
the mimosa (information that is otherwise
hidden in the ASIC) and can understand
why it behaves the way it behaves.

For this we need the full input at
defined times (e.g. between the
clock edges) and feed it into the
simulation.