Added pwm handler and unit tests
I decided to define the comparator
comparison voltages with the on-board
microcontroller. I probably want to
change the definitions of 'cold' or
'dark' and therefore, it is useful to
set it with uart commands.

The STM32G474RE mcu does not
have 6 independent DAC channels.
Therefore, I use pwm-signals fed
through a RC low pass filter to define
the comparator levels.